Finish moving repl to HIR, remove MIR and old dirs
