--- components/lark-build-hir/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "lark-build-hir"
-version = "0.1.0"
-authors = ["Jonathan Turner <jonathan.d.turner@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-lark-debug-with = { path = "../lark-debug-with" }
-lark-hir = { path = "../lark-hir" }
-lark-type-check = { path = "../lark-type-check" }
-lark-query-system = { path = "../lark-query-system" }
-lark-intern = { path = "../lark-intern" }
-lark-entity = { path = "../lark-entity" }
-lark-parser = { path = "../lark-parser" }
-lark-string = { path = "../lark-string" }
-lark-ty = { path = "../lark-ty" }
-lark-error = { path = "../lark-error" }
-tempfile = "2.0"
-cc = "1.0"
