+++ components/lark-build/src/codegen_rust.rs
@@ -1,14 +1,53 @@
 use lark_debug_with::DebugWith;
 use lark_entity::{Entity, EntityData, ItemKind, LangItem};
 use lark_error::{Diagnostic, WithError};
+use lark_hir as hir;
 use lark_intern::{Intern, Untern};
-use lark_mir::{
-    BasicBlock, BinOp, FnBytecode, MirDatabase, Operand, OperandData, Place, PlaceData, Rvalue,
-    RvalueData, Statement, StatementKind,
-};
 use lark_parser::{ParserDatabase, ParserDatabaseExt};
 use lark_query_system::LarkDatabase;
 use lark_ty::Ty;
+
+fn build_variable_name(
+    db: &LarkDatabase,
+    fn_body: &std::sync::Arc<hir::FnBody>,
+    variable: lark_hir::Variable,
+) -> String {
+    let variable_data = fn_body.tables[variable];
+    let identifier = fn_body.tables[variable_data.name];
+    identifier.text.untern(db).to_string()
+}
+
+fn build_entity_name(db: &LarkDatabase, entity: Entity) -> String {
+    let entity_data = entity.untern(db);
+    match entity_data {
+        EntityData::LangItem(LangItem::False) => "false".into(),
+        EntityData::LangItem(LangItem::True) => "true".into(),
+        EntityData::LangItem(LangItem::Debug) => "println!".into(),
+        EntityData::ItemName { id, .. } => id.untern(db).to_string(),
+        x => unimplemented!("Unsupported entity name: {:#?}", x),
+    }
+}
+
+pub fn build_place(
+    db: &LarkDatabase,
+    fn_body: &std::sync::Arc<hir::FnBody>,
+    place: hir::Place,
+) -> String {
+    match &fn_body.tables[place] {
+        hir::PlaceData::Variable(variable) => build_variable_name(db, fn_body, *variable),
+        hir::PlaceData::Entity(entity) => build_entity_name(db, *entity),
+        hir::PlaceData::Field { owner, name } => {
+            let identifier = fn_body.tables[*name];
+
+            format!(
+                "{}.{}",
+                build_place(db, fn_body, *owner),
+                identifier.text.untern(db).to_string()
+            )
+        }
+        hir::PlaceData::Temporary(expression) => build_expression(db, fn_body, *expression),
+    }
+}
 
 pub fn build_type(db: &LarkDatabase, ty: &Ty<lark_ty::declaration::Declaration>) -> String {
     let boolean_entity = EntityData::LangItem(LangItem::Boolean).intern(db);
@@ -31,66 +70,6 @@ pub fn build_type(db: &LarkDatabase, ty: &Ty<lark_ty::declaration::Declaration>)
             }
             _ => unimplemented!("Unknown base kind"),
         },
-    }
-}
-
-fn build_variable_name(
-    db: &LarkDatabase,
-    fn_bytecode: &std::sync::Arc<FnBytecode>,
-    variable: lark_mir::Variable,
-) -> String {
-    let variable_data = fn_bytecode.tables[variable];
-    let identifier = fn_bytecode.tables[variable_data.name];
-    identifier.text.untern(db).to_string()
-}
-
-fn build_entity_name(
-    db: &LarkDatabase,
-    _fn_bytecode: &std::sync::Arc<FnBytecode>,
-    entity: Entity,
-) -> String {
-    let entity_data = entity.untern(db);
-    match entity_data {
-        EntityData::LangItem(LangItem::False) => "false".into(),
-        EntityData::LangItem(LangItem::True) => "true".into(),
-        EntityData::LangItem(LangItem::Debug) => "println!".into(),
-        EntityData::ItemName { id, .. } => id.untern(db).to_string(),
-        x => unimplemented!("Unsupported entity name: {:#?}", x),
-    }
-}
-
-pub fn build_place(
-    db: &LarkDatabase,
-    fn_bytecode: &std::sync::Arc<FnBytecode>,
-    place: Place,
-) -> String {
-    match &fn_bytecode.tables[place] {
-        PlaceData::Variable(variable) => build_variable_name(db, fn_bytecode, *variable),
-        PlaceData::Entity(entity) => build_entity_name(db, fn_bytecode, *entity),
-        PlaceData::Field { owner, name } => {
-            let identifier = fn_bytecode.tables[*name];
-
-            format!(
-                "{}.{}",
-                build_place(db, fn_bytecode, *owner),
-                identifier.text.untern(db).to_string()
-            )
-        }
-    }
-}
-
-fn build_operand(
-    db: &LarkDatabase,
-    fn_bytecode: &std::sync::Arc<FnBytecode>,
-    operand: Operand,
-) -> String {
-    match &fn_bytecode.tables[operand] {
-        OperandData::ConstantInt(i) => format!("{}", i),
-        OperandData::ConstantString(s) => format!("\"{}\"", s),
-        OperandData::Copy(place) | OperandData::Move(place) => {
-            //FIXME: separate copy and move
-            build_place(db, fn_bytecode, *place)
-        }
     }
 }
 
@@ -124,121 +103,171 @@ pub fn codegen_struct(
     }
 }
 
-pub fn codegen_rvalue(
+pub fn build_expression(
     db: &LarkDatabase,
-    fn_bytecode: &std::sync::Arc<FnBytecode>,
-    rvalue: Rvalue,
-    output: &mut String,
-) {
-    match &fn_bytecode.tables[rvalue] {
-        RvalueData::Use(operand) => output.push_str(&build_operand(db, fn_bytecode, *operand)),
-        RvalueData::Call(entity, args) => {
-            output.push_str(&build_entity_name(db, fn_bytecode, *entity));
+    fn_body: &std::sync::Arc<hir::FnBody>,
+    expression: hir::Expression,
+) -> String {
+    match fn_body.tables[expression] {
+        hir::ExpressionData::Let {
+            variable,
+            initializer,
+            body,
+        } => match initializer {
+            Some(init_expression) => format!(
+                "{{ let {} = {};\n{}}}",
+                build_variable_name(db, fn_body, variable),
+                build_expression(db, fn_body, init_expression),
+                build_expression(db, fn_body, body),
+            ),
+            None => format!("let {};\n", build_variable_name(db, fn_body, variable)),
+        },
 
-            output.push_str("(");
+        hir::ExpressionData::Place { place } => build_place(db, fn_body, place),
+
+        hir::ExpressionData::Assignment { place, value } => format!(
+            "{} = {};\n",
+            build_place(db, fn_body, place),
+            build_expression(db, fn_body, value)
+        ),
+
+        hir::ExpressionData::MethodCall { method, arguments } => {
+            let mut arguments = arguments.iter(fn_body);
+            let mut output = String::new();
+
+            output.push_str(&build_expression(db, fn_body, arguments.next().unwrap()));
+
+            let method_name = fn_body.tables[method].text.untern(db);
+            output.push_str(&format!(".{}(", method_name));
+
             let mut first = true;
-
-            match entity.untern(db) {
-                EntityData::LangItem(LangItem::Debug) => {
-                    output.push_str("\"{}\"");
-                    first = false;
-                }
-                _ => {}
-            }
-
-            for arg in args.iter(fn_bytecode) {
+            for argument in arguments {
                 if !first {
                     output.push_str(", ");
                 } else {
                     first = false;
                 }
-
-                output.push_str(&build_operand(db, fn_bytecode, arg));
+                output.push_str(&build_expression(db, fn_body, argument));
             }
             output.push_str(")");
+
+            output
         }
-        RvalueData::Aggregate(entity, args) => {
-            output.push_str(&build_entity_name(db, fn_bytecode, *entity));
-            output.push_str("{");
 
-            let members = db.members(*entity).unwrap();
+        hir::ExpressionData::Call {
+            function,
+            arguments,
+        } => {
+            let mut output = String::new();
 
-            for (member, arg) in members.iter().zip(args.iter(fn_bytecode)) {
-                let member_name = member.name.untern(db);
+            output.push_str(&build_place(db, fn_body, function));
 
-                output.push_str(&format!(
-                    "{}: {},",
-                    member_name,
-                    build_operand(db, fn_bytecode, arg)
-                ));
+            output.push_str("(");
+
+            let mut first = true;
+
+            match fn_body.tables[function] {
+                hir::PlaceData::Entity(entity) => match entity.untern(db) {
+                    EntityData::LangItem(LangItem::Debug) => {
+                        output.push_str("\"{}\"");
+                        first = false;
+                    }
+                    _ => {}
+                },
+                _ => {}
             }
 
-            output.push_str("}");
+            for argument in arguments.iter(fn_body) {
+                if !first {
+                    output.push_str(", ");
+                } else {
+                    first = false;
+                }
+                output.push_str(&build_expression(db, fn_body, argument));
+            }
+            output.push_str(")");
+
+            output
         }
-        RvalueData::BinaryOp(bin_op, lhs, rhs) => match bin_op {
-            BinOp::Add => {
-                output.push_str(&format!(
-                    "({}+{})",
-                    build_operand(db, fn_bytecode, *lhs),
-                    build_operand(db, fn_bytecode, *rhs)
-                ));
-            }
-            BinOp::Sub => {
-                output.push_str(&format!(
-                    "({}-{})",
-                    build_operand(db, fn_bytecode, *lhs),
-                    build_operand(db, fn_bytecode, *rhs)
-                ));
-            }
+
+        hir::ExpressionData::Sequence { first, second } => format!(
+            "{};\n {}",
+            build_expression(db, fn_body, first),
+            build_expression(db, fn_body, second)
+        ),
+
+        hir::ExpressionData::If {
+            condition,
+            if_true,
+            if_false,
+        } => format!(
+            "if {} {{ {} \n}} else {{ {} \n}}",
+            build_expression(db, fn_body, condition),
+            build_expression(db, fn_body, if_true),
+            build_expression(db, fn_body, if_false)
+        ),
+
+        hir::ExpressionData::Binary {
+            operator,
+            left,
+            right,
+        } => format!(
+            "({} {} {})",
+            build_expression(db, fn_body, left),
+            match operator {
+                hir::BinaryOperator::Add => "+",
+                hir::BinaryOperator::Subtract => "-",
+                hir::BinaryOperator::Multiply => "*",
+                hir::BinaryOperator::Divide => "/",
+                hir::BinaryOperator::Equals => "==",
+                hir::BinaryOperator::NotEquals => "!=",
+            },
+            build_expression(db, fn_body, right),
+        ),
+
+        hir::ExpressionData::Unary { operator, value } => format!(
+            "{}({})",
+            match operator {
+                hir::UnaryOperator::Not => "!",
+            },
+            build_expression(db, fn_body, value)
+        ),
+
+        hir::ExpressionData::Literal { data } => match data {
+            hir::LiteralData {
+                kind: hir::LiteralKind::String,
+                value,
+            } => format!("\"{}\"", value.untern(db)),
+            hir::LiteralData {
+                kind: hir::LiteralKind::UnsignedInteger,
+                value,
+            } => format!("{}", value.untern(db)),
         },
-    }
-}
 
-pub fn codegen_statement(
-    db: &LarkDatabase,
-    statement: Statement,
-    fn_bytecode: &std::sync::Arc<FnBytecode>,
-    output: &mut String,
-) {
-    let statement_data = &fn_bytecode.tables[statement];
+        hir::ExpressionData::Unit {} => "()".to_string(),
 
-    match &statement_data.kind {
-        StatementKind::Expression(rvalue) => {
-            codegen_rvalue(db, fn_bytecode, *rvalue, output);
-        }
-        StatementKind::Assign(lvalue, rvalue) => {
-            output.push_str(&format!("{} = ", build_place(db, fn_bytecode, *lvalue)));
-            codegen_rvalue(db, fn_bytecode, *rvalue, output);
-        }
-        StatementKind::StorageLive(variable) => {
-            output.push_str(&format!(
-                "{{\nlet {};\n",
-                build_variable_name(db, fn_bytecode, *variable)
-            ));
-        }
-        StatementKind::StorageDead(_) => {
-            output.push_str("\n}\n");
-        }
-    }
-}
+        hir::ExpressionData::Aggregate { entity, fields } => {
+            let mut output = String::new();
 
-pub fn codegen_basic_block(
-    db: &LarkDatabase,
-    fn_bytecode: &std::sync::Arc<FnBytecode>,
-    basic_block: BasicBlock,
-    output: &mut String,
-) {
-    let basic_block_data = &fn_bytecode.tables[basic_block];
-
-    let mut first = true;
-
-    for statement in basic_block_data.statements.iter(&fn_bytecode) {
-        if !first {
-            output.push_str(";\n")
-        } else {
-            first = false;
+            output.push_str(&build_entity_name(db, entity));
+            output.push_str("{");
+            for field in fields.iter(fn_body) {
+                let identified_expression = fn_body.tables[field];
+                output.push_str(&format!(
+                    "{}: {}",
+                    fn_body.tables[identified_expression.identifier]
+                        .text
+                        .untern(db),
+                    build_expression(db, fn_body, identified_expression.expression),
+                ));
+            }
+            output.push_str("}");
+            output
         }
-        codegen_statement(db, statement, fn_bytecode, output);
+
+        hir::ExpressionData::Error { .. } => {
+            panic!("Can not codegen in the presence of errors");
+        }
     }
 }
 
@@ -250,24 +279,23 @@ pub fn codegen_function(
     let mut output = String::new();
     let mut errors: Vec<Diagnostic> = vec![];
 
-    let fn_bytecode = db.fn_bytecode(entity).accumulate_errors_into(&mut errors);
+    let fn_body = db.fn_body(entity).accumulate_errors_into(&mut errors);
+
     let signature = db
         .signature(entity)
         .accumulate_errors_into(&mut errors)
         .unwrap();
+
+    let arguments = fn_body.arguments.unwrap();
 
     let name = id.untern(db);
 
     output.push_str(&format!("fn {}(", name));
 
     let mut first = true;
-    for (argument, argument_type) in fn_bytecode
-        .arguments
-        .iter(&fn_bytecode)
-        .zip(signature.inputs.iter())
-    {
-        let variable_data = fn_bytecode.tables[argument];
-        let identifier = fn_bytecode.tables[variable_data.name];
+    for (argument, argument_type) in arguments.iter(&fn_body).zip(signature.inputs.iter()) {
+        let variable_data = fn_body.tables[argument];
+        let identifier = fn_body.tables[variable_data.name];
 
         let argument_name = identifier.text.untern(db);
 
@@ -283,11 +311,10 @@ pub fn codegen_function(
 
     output.push_str(") -> ");
     output.push_str(&format!("{}", build_type(db, &signature.output)));
-    output.push_str(" {\n");
-    for basic_block in fn_bytecode.basic_blocks.iter(&fn_bytecode) {
-        codegen_basic_block(db, &fn_bytecode, basic_block, &mut output);
-    }
-    output.push_str("}\n");
+    output.push_str(&format!(
+        " {{\n{} }}\n",
+        build_expression(db, &fn_body, fn_body.root_expression)
+    ));
 
     WithError {
         value: output,
