+++ components/lark-cli/src/lib.rs
@@ -22,57 +22,6 @@ mod ide;
 mod repl;
 mod run;
 
-pub fn codegen(file_name: &str) {
-    use flexi_logger::{opt_format, Logger};
-    use language_reporting::{emit, Diagnostic, Label, Severity};
-    use languageserver_types::Position;
-    use lark_entity::{EntityData, ItemKind, MemberKind};
-    use lark_intern::{Intern, Untern};
-    use lark_language_server::{lsp_serve, LspResponder};
-    use lark_mir::MirDatabase;
-    use lark_parser::{ParserDatabase, ParserDatabaseExt};
-    use lark_query_system::ls_ops::Cancelled;
-    use lark_query_system::ls_ops::LsDatabase;
-    use lark_query_system::LarkDatabase;
-    use lark_query_system::QuerySystem;
-    use lark_span::{ByteIndex, FileName, IntoFileName, Span};
-    use lark_task_manager::Actor;
-    use salsa::Database;
-    use std::borrow::Cow;
-    use std::collections::HashMap;
-    use std::fs::File;
-    use std::io::Read;
-    use std::sync::Arc;
-    use std::{env, io};
-    use termcolor::{ColorChoice, StandardStream, WriteColor};
-
-    let mut file = match File::open(file_name) {
-        Ok(f) => f,
-        Err(err) => {
-            eprintln!("failed to open `{}`: {}", file_name, err);
-            return;
-        }
-    };
-
-    let mut contents = String::new();
-    match file.read_to_string(&mut contents) {
-        Ok(_bytes_read) => {}
-        Err(err) => {
-            eprintln!("failed to read `{}`: {}", file_name, err);
-            return;
-        }
-    }
-
-    let mut db = LarkDatabase::default();
-
-    let file_id: FileName = file_name.into_file_name(&db);
-    db.add_file(file_id, contents);
-
-    let rust_file = lark_build_hir::codegen(&db, lark_build_hir::CodegenType::Rust);
-
-    println!("{}", rust_file.into_value());
-}
-
 pub fn main() {
     Logger::with_env_or_str("error,lark_query_system=info")
         .log_to_file()
@@ -84,7 +33,6 @@ pub fn main() {
     let mut args = std::env::args();
 
     match (args.next(), args.next(), args.next(), args.next()) {
-        (_, Some(ref cmd), Some(ref x), None) if cmd == "codegen" => codegen(x),
         (_, Some(ref cmd), Some(ref x), Some(ref out)) if cmd == "build" => {
             build::build(x, Some(out))
         }
