+++ components/lark-cli/src/repl.rs
@@ -1,8 +1,9 @@
 use crate::build::LarkDatabaseExt;
+use lark_debug_with::DebugWith;
 use lark_entity::{EntityData, ItemKind};
 use lark_eval::Value;
+use lark_hir as hir;
 use lark_intern::{Intern, Untern};
-use lark_mir::{FnBytecode, MirDatabase, StatementKind, Variable};
 use lark_parser::{ParserDatabase, ParserDatabaseExt};
 use lark_query_system::ls_ops::LsDatabase;
 use lark_query_system::LarkDatabase;
@@ -12,9 +13,7 @@ use termcolor::{ColorChoice, StandardStream, WriteColor};
 
 const REPL_FILENAME: &str = "__REPL__.lark";
 
-pub fn get_bytecode(
-    db: &LarkDatabase,
-) -> lark_error::WithError<std::sync::Arc<lark_mir::FnBytecode>> {
+pub fn get_body(db: &LarkDatabase) -> lark_error::WithError<std::sync::Arc<lark_hir::FnBody>> {
     let main_name = "main".intern(&db);
     let repl_filename = REPL_FILENAME.intern(&db);
     let entities = db.top_level_entities_in_file(repl_filename);
@@ -27,8 +26,8 @@ pub fn get_bytecode(
                 ..
             } => {
                 if id == main_name {
-                    let bytecode = db.fn_bytecode(entity);
-                    return bytecode;
+                    let body = db.fn_body(entity);
+                    return body;
                 }
             }
             _ => {}
@@ -39,11 +38,17 @@ pub fn get_bytecode(
 }
 
 pub fn repl() {
-    let mut fn_body: Vec<String> = vec![];
+    let mut virtual_fn: Vec<String> = vec![];
     let mut io_handler = lark_eval::IOHandler::new(false);
     let mut db = LarkDatabase::default();
-    let mut variables: HashMap<Variable, Vec<Value>> = HashMap::new();
-    let mut num_to_skip = 0;
+
+    let _ = db.add_file(
+        REPL_FILENAME,
+        format!("def main() {{\n{}\n}}", virtual_fn.join("\n")),
+    );
+
+    let mut eval_state = lark_eval::EvalState::new();
+    eval_state.is_repl = true;
 
     println!("Lark repl (:? - command help)");
     loop {
@@ -60,12 +65,12 @@ pub fn repl() {
         if input == ":p" {
             println!(
                 "Will execute: {}",
-                format!("def main() {{\n{}\n}}", fn_body.join("\n"))
+                format!("def main() {{\n{}\n}}", virtual_fn.join("\n"))
             );
             continue;
         }
         if input == ":v" {
-            println!("{:#?}", variables);
+            println!("{:#?}", eval_state.variables);
             continue;
         }
         if input == ":?" {
@@ -76,11 +81,11 @@ pub fn repl() {
             continue;
         }
 
-        fn_body.push(input);
+        virtual_fn.push(input);
 
         let _ = db.add_file(
             REPL_FILENAME,
-            format!("def main() {{\n{}\n}}", fn_body.join("\n")),
+            format!("def main() {{\n{}\n}}", virtual_fn.join("\n")),
         );
 
         let writer = StandardStream::stderr(ColorChoice::Auto);
@@ -90,51 +95,21 @@ pub fn repl() {
 
         if error_count > 0 {
             // The last command was bad. Let's remove it from our function body
-            fn_body.pop();
+            virtual_fn.pop();
         } else {
             // No errors, so let's run the last line of our function body
-            let fn_bytecode = get_bytecode(&mut db).value;
+            let fn_body = get_body(&mut db).value;
 
-            //println!("{:#?}", fn_bytecode);
+            let output = lark_eval::eval_expression(
+                &db,
+                &fn_body,
+                fn_body.root_expression,
+                &mut eval_state,
+                &mut io_handler,
+            );
 
-            let mut num_to_skip_remaining = num_to_skip;
-            let mut num_to_skip_next = 0;
-            let mut output = Value::Void;
+            eval_state.skip_until = eval_state.current_expression;
 
-            for basic_block in fn_bytecode.basic_blocks.iter(&fn_bytecode) {
-                let basic_block_data = &fn_bytecode.tables[basic_block];
-
-                for statement in basic_block_data.statements.iter(&fn_bytecode) {
-                    if num_to_skip_remaining > 0 {
-                        num_to_skip_remaining -= 1;
-                    } else {
-                        let statement_data = &fn_bytecode.tables[statement];
-
-                        match &statement_data.kind {
-                            StatementKind::StorageLive(variable) => {
-                                num_to_skip_next += 1;
-                                // Because we manage our own variables, we have to do a bit of bookkeeping
-                                lark_eval::create_variable(&mut variables, *variable);
-                            }
-                            StatementKind::StorageDead(_) => {
-                                // In the repl, we don't currently delete old variables
-                            }
-                            _ => {
-                                num_to_skip_next += 1;
-                                output = lark_eval::eval_statement(
-                                    &mut db,
-                                    &fn_bytecode,
-                                    statement,
-                                    &mut variables,
-                                    &mut io_handler,
-                                );
-                            }
-                        }
-                    }
-                }
-            }
-
-            num_to_skip += num_to_skip_next;
             match output {
                 Value::Void => {}
                 x => println!("{}", x),
