--- components/lark-eval-hir/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "lark-eval-hir"
-version = "0.1.0"
-authors = ["Jonathan Turner <jonathan.d.turner@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-lark-hir = { path = "../lark-hir" }
-lark-type-check = { path = "../lark-type-check" }
-lark-query-system = { path = "../lark-query-system" }
-lark-entity = { path = "../lark-entity" }
-lark-parser = { path = "../lark-parser" }
-lark-span = { path = "../lark-span" }
-lark-string = { path = "../lark-string" }
-lark-ty = { path = "../lark-ty" }
-lark-error = { path = "../lark-error" }
-lark-intern = { path = "../lark-intern" }
-lark-debug-with = { path = "../lark-debug-with" }
