+++ components/lark-eval/Cargo.toml
@@ -5,7 +5,6 @@ authors = ["Jonathan Turner <jonathan.d.turner@gmail.com>"]
 edition = "2018"
 
 [dependencies]
-lark-mir = { path = "../lark-mir" }
 lark-hir = { path = "../lark-hir" }
 lark-type-check = { path = "../lark-type-check" }
 lark-query-system = { path = "../lark-query-system" }
