+++ components/lark-eval/src/lib.rs
@@ -1,39 +1,58 @@
 use lark_debug_with::DebugWith;
 use lark_entity::{EntityData, ItemKind, LangItem};
+use lark_hir as hir;
 use lark_intern::{Intern, Untern};
-use lark_mir::{
-    BasicBlock, BinOp, FnBytecode, IdentifierData, MirDatabase, Operand, OperandData, Place,
-    PlaceData, Rvalue, RvalueData, Statement, StatementKind, Variable,
-};
 use lark_parser::{ParserDatabase, ParserDatabaseExt};
 use lark_query_system::LarkDatabase;
 use std::collections::HashMap;
 use std::fmt;
 
-#[derive(Clone, Debug)]
-pub enum Value {
-    Void,
-    Bool(bool),
-    U32(u32),
-    Str(String),
-    Struct(HashMap<lark_string::GlobalIdentifier, Value>),
-    Reference(usize), // a reference into the value stack
+pub struct EvalState {
+    pub variables: HashMap<hir::Variable, Vec<Value>>,
+    pub skip_until: Option<hir::Expression>,
+    pub current_expression: Option<hir::Expression>,
+    pub is_repl: bool,
 }
 
-impl fmt::Display for Value {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "{}",
-            match self {
-                Value::U32(u) => u.to_string(),
-                Value::Str(s) => s.clone(),
-                Value::Bool(b) => b.to_string(),
-                Value::Reference(r) => format!("reference to {}", r),
-                Value::Void => "<void>".into(),
-                Value::Struct(s) => format!("{:?}", s),
+impl EvalState {
+    pub fn create_variable(&mut self, variable: hir::Variable) {
+        let variable_stack = self.variables.entry(variable).or_insert(Vec::new());
+        variable_stack.push(Value::Void);
+    }
+
+    pub fn pop_variable(&mut self, variable: hir::Variable) {
+        let variable_stack = self.variables.get_mut(&variable).unwrap();
+        variable_stack.pop();
+    }
+
+    pub fn assign_to_variable(&mut self, variable: hir::Variable, value: Value) {
+        let variable_stack = self.variables.get_mut(&variable).unwrap();
+        *variable_stack.last_mut().unwrap() = value;
+    }
+
+    pub fn new() -> EvalState {
+        EvalState {
+            variables: HashMap::new(),
+            skip_until: None,
+            current_expression: None,
+            is_repl: false,
+        }
+    }
+
+    pub fn set_current_expression(&mut self, expression: hir::Expression) {
+        self.current_expression = Some(expression);
+
+        if self.skip_until.is_some() {
+            if let Some(skip_until_expr) = self.skip_until {
+                if skip_until_expr == expression {
+                    self.skip_until = None;
+                }
             }
-        )
+        }
+    }
+
+    pub fn ready_to_execute(&self) -> bool {
+        self.skip_until.is_none()
     }
 }
 
@@ -62,221 +81,264 @@ impl IOHandler {
     }
 }
 
+#[derive(Clone, Debug)]
+pub enum Value {
+    Void,
+    Bool(bool),
+    U32(u32),
+    Str(String),
+    Struct(HashMap<lark_string::GlobalIdentifier, Value>),
+    Reference(usize), // a reference into the value stack
+}
+
+impl fmt::Display for Value {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(
+            f,
+            "{}",
+            match self {
+                Value::U32(u) => u.to_string(),
+                Value::Str(s) => s.clone(),
+                Value::Bool(b) => b.to_string(),
+                Value::Reference(r) => format!("reference to {}", r),
+                Value::Void => "<void>".into(),
+                Value::Struct(s) => format!("{:?}", s),
+            }
+        )
+    }
+}
+
 pub fn eval_place(
     db: &LarkDatabase,
-    fn_bytecode: &FnBytecode,
-    place: Place,
-    variables: &mut HashMap<Variable, Vec<Value>>,
+    fn_body: &hir::FnBody,
+    place: hir::Place,
+    state: &mut EvalState,
 ) -> Value {
-    let place_data = &fn_bytecode.tables[place];
+    let place_data = &fn_body.tables[place];
 
     match place_data {
-        PlaceData::Entity(entity) => match entity.untern(db) {
+        hir::PlaceData::Entity(entity) => match entity.untern(db) {
             EntityData::LangItem(LangItem::True) => Value::Bool(true),
             EntityData::LangItem(LangItem::False) => Value::Bool(false),
             _ => unimplemented!("EntityData not yet support in eval"),
         },
-        PlaceData::Variable(variable) => {
-            let stack = variables.get(variable).unwrap();
+        hir::PlaceData::Variable(variable) => {
+            let stack = state.variables.get(variable).unwrap();
             stack.last().unwrap().clone()
         }
-        PlaceData::Field { owner, name } => {
-            let target = eval_place(db, fn_bytecode, *owner, variables);
+        hir::PlaceData::Field { owner, name } => {
+            let target = eval_place(db, fn_body, *owner, state);
             match target {
-                Value::Struct(s) => match fn_bytecode.tables[*name] {
-                    IdentifierData { text } => s.get(&text).unwrap().clone(),
+                Value::Struct(s) => match fn_body.tables[*name] {
+                    hir::IdentifierData { text } => s.get(&text).unwrap().clone(),
                 },
                 _ => panic!("Member access (.) into value that is not a struct"),
             }
         }
+        hir::PlaceData::Temporary { .. } => unimplemented!("Can't yet eval temporary places"),
     }
 }
 
-pub fn eval_operand(
+pub fn eval_expression(
     db: &LarkDatabase,
-    fn_bytecode: &FnBytecode,
-    operand: Operand,
-    variables: &mut HashMap<Variable, Vec<Value>>,
-) -> Value {
-    let operand_data = &fn_bytecode.tables[operand];
-
-    match operand_data {
-        OperandData::Copy(place) | OperandData::Move(place) => {
-            eval_place(db, fn_bytecode, *place, variables)
-        }
-        OperandData::ConstantInt(u) => Value::U32(*u),
-        _ => unimplemented!("Operand not yet supported"),
-    }
-}
-
-pub fn eval_rvalue(
-    db: &LarkDatabase,
-    fn_bytecode: &FnBytecode,
-    rvalue: Rvalue,
-    variables: &mut HashMap<Variable, Vec<Value>>,
+    fn_body: &hir::FnBody,
+    expression: hir::Expression,
+    state: &mut EvalState,
     io_handler: &mut IOHandler,
 ) -> Value {
-    let rvalue_data = &fn_bytecode.tables[rvalue];
+    match fn_body.tables[expression] {
+        hir::ExpressionData::Unit { .. } => {}
+        _ => state.set_current_expression(expression),
+    }
 
-    match rvalue_data {
-        RvalueData::Use(operand) => eval_operand(db, fn_bytecode, *operand, variables),
-        RvalueData::Call(entity, operands) => match entity.untern(db) {
-            EntityData::LangItem(LangItem::Debug) => {
-                for operand in operands.iter(fn_bytecode) {
-                    let result = eval_operand(db, fn_bytecode, operand, variables);
-                    io_handler.println(format!("{}", result));
+    let ready_to_execute = state.ready_to_execute();
+
+    match fn_body.tables[expression] {
+        hir::ExpressionData::Let {
+            variable,
+            initializer,
+            body,
+        } => {
+            if ready_to_execute {
+                state.create_variable(variable);
+            }
+
+            if let Some(expression) = initializer {
+                let result = eval_expression(db, fn_body, expression, state, io_handler);
+
+                if ready_to_execute {
+                    state.assign_to_variable(variable, result);
                 }
+            }
 
+            let body_result = eval_expression(db, fn_body, body, state, io_handler);
+
+            if !state.is_repl {
+                state.pop_variable(variable);
+            }
+
+            body_result
+        }
+
+        hir::ExpressionData::Place { place } => {
+            if ready_to_execute {
+                eval_place(db, fn_body, place, state)
+            } else {
                 Value::Void
             }
-            EntityData::ItemName { .. } => {
-                let bytecode = db.fn_bytecode(*entity).value;
+        }
 
-                for (arg, param) in operands
-                    .iter(fn_bytecode)
-                    .zip(bytecode.arguments.iter(&bytecode))
-                {
-                    let arg_value = eval_operand(db, fn_bytecode, arg, variables);
-                    create_variable(variables, param);
-                    assign_to_variable(variables, param, arg_value);
+        hir::ExpressionData::Assignment { place, value } => {
+            let rhs = eval_expression(db, fn_body, value, state, io_handler);
+            match &fn_body.tables[place] {
+                hir::PlaceData::Variable(variable) => {
+                    if ready_to_execute {
+                        println!("Assigning {:?} to variable", rhs);
+                        state.assign_to_variable(*variable, rhs);
+                    }
                 }
-
-                let return_value = eval_function(db, &bytecode, variables, io_handler);
-
-                for argument in fn_bytecode.arguments.iter(&fn_bytecode) {
-                    pop_variable(variables, argument);
-                }
-
-                return_value
-            }
-            x => unimplemented!(
-                "EntityData not yet supported in eval: {:#?}",
-                x.debug_with(db)
-            ),
-        },
-        RvalueData::Aggregate(entity, args) => {
-            let members = db.members(*entity).unwrap();
-            let mut result_struct = HashMap::new();
-
-            for (member, arg) in members.iter().zip(args.iter(fn_bytecode)) {
-                let arg_result = eval_operand(db, fn_bytecode, arg, variables);
-                result_struct.insert(member.name, arg_result);
-            }
-
-            Value::Struct(result_struct)
-        }
-        RvalueData::BinaryOp(binop, lhs, rhs) => {
-            let lhs_eval = eval_operand(db, fn_bytecode, *lhs, variables);
-            let rhs_eval = eval_operand(db, fn_bytecode, *rhs, variables);
-
-            match binop {
-                BinOp::Add => match (lhs_eval, rhs_eval) {
-                    (Value::U32(l), Value::U32(r)) => Value::U32(l + r),
-                    _ => panic!("Addition of non-numeric values"),
-                },
-                BinOp::Sub => match (lhs_eval, rhs_eval) {
-                    (Value::U32(l), Value::U32(r)) => Value::U32(l - r),
-                    _ => panic!("Subtraction of non-numeric values"),
-                },
-            }
-        }
-    }
-}
-
-pub fn create_variable(variables: &mut HashMap<Variable, Vec<Value>>, variable: Variable) {
-    let variable_stack = variables.entry(variable).or_insert(Vec::new());
-    variable_stack.push(Value::Void);
-}
-
-pub fn pop_variable(variables: &mut HashMap<Variable, Vec<Value>>, variable: Variable) {
-    let variable_stack = variables.get_mut(&variable).unwrap();
-    variable_stack.pop();
-}
-
-pub fn assign_to_variable(
-    variables: &mut HashMap<Variable, Vec<Value>>,
-    variable: Variable,
-    value: Value,
-) {
-    let variable_stack = variables.get_mut(&variable).unwrap();
-    *variable_stack.last_mut().unwrap() = value;
-}
-
-pub fn eval_statement(
-    db: &LarkDatabase,
-    fn_bytecode: &FnBytecode,
-    statement: Statement,
-    variables: &mut HashMap<Variable, Vec<Value>>,
-    io_handler: &mut IOHandler,
-) -> Value {
-    let statement_data = &fn_bytecode.tables[statement];
-
-    match &statement_data.kind {
-        StatementKind::Expression(rvalue) => {
-            eval_rvalue(db, fn_bytecode, *rvalue, variables, io_handler)
-        }
-        StatementKind::Assign(place, rvalue) => {
-            let rhs = eval_rvalue(db, fn_bytecode, *rvalue, variables, io_handler);
-            match &fn_bytecode.tables[*place] {
-                PlaceData::Variable(variable) => assign_to_variable(variables, *variable, rhs),
                 _ => unimplemented!("PlaceData not yet supported in eval"),
             }
             Value::Void
         }
-        _ => {
-            // If we get here, something went wrong. Statements like StorageLive/StorageDead
-            // Need to get handled where they won't effect the return value of the block
-            panic!("Unexpected StatementKind");
+
+        hir::ExpressionData::Call {
+            function,
+            arguments,
+        } => match &fn_body.tables[function] {
+            hir::PlaceData::Entity(entity) => match entity.untern(db) {
+                EntityData::LangItem(LangItem::Debug) => {
+                    for argument in arguments.iter(fn_body) {
+                        let result = eval_expression(db, fn_body, argument, state, io_handler);
+
+                        if ready_to_execute {
+                            io_handler.println(format!("{}", result));
+                        }
+                    }
+
+                    Value::Void
+                }
+                EntityData::ItemName { .. } => {
+                    let target = db.fn_body(*entity).value;
+
+                    for (arg, param) in arguments
+                        .iter(fn_body)
+                        .zip(target.arguments.unwrap().iter(&target))
+                    {
+                        let arg_value = eval_expression(db, fn_body, arg, state, io_handler);
+                        state.create_variable(param);
+                        state.assign_to_variable(param, arg_value);
+                    }
+
+                    let return_value = if ready_to_execute {
+                        eval_function(db, &target, state, io_handler)
+                    } else {
+                        Value::Void
+                    };
+
+                    for argument in target.arguments.unwrap().iter(&target) {
+                        state.pop_variable(argument);
+                    }
+
+                    return_value
+                }
+                x => unimplemented!(
+                    "Call entity not yet supported in eval: {:#?}",
+                    x.debug_with(db)
+                ),
+            },
+            x => unimplemented!("Call not yet supported in eval: {:#?}", x.debug_with(db)),
+        },
+
+        hir::ExpressionData::Sequence { first, second } => {
+            eval_expression(db, fn_body, first, state, io_handler);
+            eval_expression(db, fn_body, second, state, io_handler)
         }
-    }
-}
 
-pub fn eval_basic_block(
-    db: &LarkDatabase,
-    fn_bytecode: &FnBytecode,
-    basic_block: BasicBlock,
-    variables: &mut HashMap<Variable, Vec<Value>>,
-    io_handler: &mut IOHandler,
-) -> Value {
-    let basic_block_data = &fn_bytecode.tables[basic_block];
-    let mut return_value = Value::Void;
+        hir::ExpressionData::Binary {
+            operator,
+            left,
+            right,
+        } => {
+            let lhs_eval = eval_expression(db, fn_body, left, state, io_handler);
+            let rhs_eval = eval_expression(db, fn_body, right, state, io_handler);
 
-    for statement in basic_block_data.statements.iter(&fn_bytecode) {
-        let statement_data = &fn_bytecode.tables[statement];
-
-        match &statement_data.kind {
-            StatementKind::StorageLive(variable) => {
-                create_variable(variables, *variable);
+            if ready_to_execute {
+                match operator {
+                    hir::BinaryOperator::Add => match (lhs_eval, rhs_eval) {
+                        (Value::U32(l), Value::U32(r)) => Value::U32(l + r),
+                        _ => panic!("Addition of non-numeric values"),
+                    },
+                    hir::BinaryOperator::Subtract => match (lhs_eval, rhs_eval) {
+                        (Value::U32(l), Value::U32(r)) => Value::U32(l - r),
+                        _ => panic!("Subtraction of non-numeric values"),
+                    },
+                    _ => unimplemented!("Operator not yet supported"),
+                }
+            } else {
+                Value::Void
             }
-            StatementKind::StorageDead(variable) => {
-                pop_variable(variables, *variable);
-            }
-            _ => return_value = eval_statement(db, fn_bytecode, statement, variables, io_handler),
         }
-    }
 
-    return_value
+        hir::ExpressionData::Literal { data } => match data {
+            hir::LiteralData {
+                kind: hir::LiteralKind::UnsignedInteger,
+                value,
+            } => {
+                if ready_to_execute {
+                    let string = value.untern(db);
+                    let value: u32 = string.parse().unwrap();
+                    Value::U32(value)
+                } else {
+                    Value::Void
+                }
+            }
+            _ => unimplemented!("Unsupported literal value"),
+        },
+
+        hir::ExpressionData::Aggregate { fields, .. } => {
+            let mut result_struct = HashMap::new();
+
+            for identified_expression in fields.iter(fn_body) {
+                let hir::IdentifiedExpressionData {
+                    identifier,
+                    expression,
+                } = fn_body.tables[identified_expression];
+                let arg_result = eval_expression(db, fn_body, expression, state, io_handler);
+
+                result_struct.insert(fn_body.tables[identifier].text, arg_result);
+            }
+
+            if ready_to_execute {
+                Value::Struct(result_struct)
+            } else {
+                Value::Void
+            }
+        }
+
+        hir::ExpressionData::Unit {} => Value::Void,
+
+        ref x => unimplemented!(
+            "Eval does not yet support this expression type: {:#?}",
+            x.debug_with(db)
+        ),
+    }
 }
 
 pub fn eval_function(
     db: &LarkDatabase,
-    fn_bytecode: &FnBytecode,
-    variables: &mut HashMap<Variable, Vec<Value>>,
+    fn_body: &hir::FnBody,
+    state: &mut EvalState,
     io_handler: &mut IOHandler,
 ) -> Value {
-    let mut return_value = Value::Void;
-    for basic_block in fn_bytecode.basic_blocks.iter(fn_bytecode) {
-        return_value = eval_basic_block(db, fn_bytecode, basic_block, variables, io_handler);
-    }
-
-    return_value
+    eval_expression(db, fn_body, fn_body.root_expression, state, io_handler)
 }
 
 pub fn eval(db: &LarkDatabase, io_handler: &mut IOHandler) {
     let input_files = db.file_names();
-    //let mut errors: Vec<Diagnostic> = vec![];
 
-    let mut variables: HashMap<Variable, Vec<Value>> = HashMap::new();
+    let mut eval_state = EvalState::new();
+
     let main_name = "main".intern(&db);
 
     for &input_file in &*input_files {
@@ -290,9 +352,9 @@ pub fn eval(db: &LarkDatabase, io_handler: &mut IOHandler) {
                     ..
                 } => {
                     if id == main_name {
-                        let bytecode = db.fn_bytecode(entity);
+                        let fn_body = db.fn_body(entity);
 
-                        eval_function(db, &bytecode.value, &mut variables, io_handler);
+                        eval_function(db, &fn_body.value, &mut eval_state, io_handler);
                     }
                 }
                 _ => {}
