--- components/lark-mir/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "lark-mir"
-version = "0.1.0"
-authors = ["Jonathan Turner <jonathan.d.turner@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-salsa = "0.8"
-
-lark-collections = { path = "../lark-collections" }
-lark-debug-derive = { path = "../lark-debug-derive" }
-lark-debug-with = { path = "../lark-debug-with" }
-lark-indices = { path = "../lark-indices" }
-lark-intern = { path = "../lark-intern" }
-lark-entity = { path = "../lark-entity" }
-lark-error = { path = "../lark-error" }
-lark-parser = { path = "../lark-parser" }
-lark-hir = { path = "../lark-hir" }
-lark-span = { path = "../lark-span" }
-lark-type-check = { path = "../lark-type-check" }
-lark-ty = { path = "../lark-ty" }
-lark-string = {path = "../lark-string" }
